requirements.txt
streamlit
pandas
mysql-connector-python
sqlalchemy
bcrypt
cryptography
python-dotenv